Use `off_folksonomy` as package name, switch to uv, refactor container image
